Integrate signup login with the database
