+++ app/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from app.products.views import product
 from app.auth.views import auths
-from app.sales.views import sale_bp
+# from app.sales.views import sale_bp
 # from app import database
 from instance.config import app_config
 import sys
@@ -19,5 +19,5 @@ def create_app(config_name):
     #Register Blueprints
     app.register_blueprint(product, url_prefix='/api/v1' )
     app.register_blueprint(auths,url_prefix='/api/v1')
-    app.register_blueprint(sale_bp,url_prefix='/api/v1')
+    # app.register_blueprint(sale_bp,url_prefix='/api/v1')
     return app
