+++ app/auth/models.py
@@ -1,10 +1,13 @@
 from app.auth.database import Database
+from flask import current_app as app
+
 
 
 #Store attendant class model
 class User(Database):
     def __init__(self,username,email,password,admin_status=False):
         Database.__init__(self,'postgres://postgres:psql@localhost:5432/store')
+
         """
         Class for creating the store attendant object
 
@@ -20,17 +23,11 @@ class User(Database):
         .format(self.username, self.email, self.password, self.admin_status))
         self.c.execute(sql)
         return True
-    
-    def select_users(self):
-        sql = ("""SELECT * from user_table """)
-        self.c.execute(sql)
-        rows = self.c.fetchall()
-        return rows
 
-    def select_a_user(self,user_id_in):
-        sql = ("""SELECT * from user_table WHERE userid = {} """.format(user_id_in))
-        self.c.execute(sql)
-        row = self.c.fetchone()
-        return row
+   
 
 
+# me = User('sanya','ken',123)
+# me.create_tables()
+# me.insert_user()
+# print(me.select_users())
