+++ app/auth/views.py
@@ -1,5 +1,6 @@
 from flask import Blueprint,request,jsonify,json
 from app.auth.models import User
+from app.auth.database import Database,db
 from werkzeug.security import check_password_hash,generate_password_hash
 
 
@@ -9,10 +10,9 @@ from werkzeug.security import check_password_hash,generate_password_hash
 # blueprint will handle all app user routes
 auths = Blueprint('auths',__name__)
 
-user_db = [] #List to hold user data
-admin_db = [] #List to hold admin data
 
 
+#create a store attendant route
 @auths.route('/users', methods=['POST'])
 def create_store_attendant():
     """
@@ -20,116 +20,51 @@ def create_store_attendant():
     If data input is not valid function will return a customise error message
 
     """
-    try:
-        data = request.data
-        data = json.loads(data)
-        user_name = data['user_name']
-        user_password = str(data['user_password'])
-        #check if content type is application/json
-        if not request.content_type == 'application/json': 
-            return jsonify({'error':'Wrong content-type'}),400
-        if user_name == "" or user_password == "":
-            return jsonify({'error':'username or password cannot be empty'}),400
-        if not isinstance(user_name,str):
-            return jsonify({'error':'username must be a string'}),400
-        if (' ' in user_name) == True:
-            return jsonify({'Error':'user name cannot contain a space'}),400
-        # check_user(user_name,user_password)
-        for fetch_user in user_db:
-            if fetch_user['user_password'] == user_password or fetch_user['user_name'] == user_name:
-                return jsonify({'error':'user already exists'}),400
-        usr_password = generate_password_hash(user_password, method='sha256')
-        #Initialise User object to add provided data
-        store_attendant = User(user_name,usr_password)
-        user_db.append(store_attendant.to_dict())
-        return jsonify({'message':'Account was successfuly created'}),201
-    except Exception:
-        return jsonify({'error':'required field/s missing'}), 400  
-
-
-@auths.route('users/admin', methods=['POST'])  
-def create_admin():
-    """
-    Function creates an admin account given that the data input is valid
-    If data input is not valid function will return a customise error message
-
-    """
-    try:
-        admin_details = json.loads(request.data)
-        admin_name = admin_details['admin_name']
-        admin_password = str(admin_details['admin_password'])
-        #check if content type is application/json
-        if not request.content_type == 'application/json': 
-            return jsonify({'error':'Wrong content-type'}),400
-        if not admin_name or not admin_password:
-            return jsonify({'error':'username or password cannot be empty'}),400
-        if not isinstance(admin_name,str):
-            return jsonify({'error':'username must be a string'}),400
-        if (' ' in admin_name) == True:
-            return jsonify({'Error':'user name cannot contain a space'}),400
-        # check_user(admin_name, admin_password)
-        for get_admin in admin_db:
-            if admin_name == get_admin['user_name']:
-                return jsonify({'error':'Admin exists'}),400
-        adm_password = generate_password_hash(admin_password, method='sha256')
-        #Initialise admin object to add provided data
-        admin = Admin(admin_name,adm_password)
-        admin_db.append(admin.to_dict())
-        return jsonify({'message':'Account was successfuly created'}),201
-    except Exception:
-        return jsonify({'error':'required field/s missing'}), 400  
-
-
+    data = request.data
+    data = json.loads(data)
+    user_name = data['user_name']
+    user_email = data['user_email']
+    user_password = str(data['user_password'])
+    #check if content type is application/json
+    if not request.content_type == 'application/json': 
+        return jsonify({'error':'Wrong content-type'}),400
+    if user_name == "" or user_password == "":
+        return jsonify({'error':'username or password cannot be empty'}),400
+    if not isinstance(user_name,str):
+        return jsonify({'error':'username must be a string'}),400
+    # if (' ' in user_name) == True:
+    #     return jsonify({'Error':'user name cannot contain a space'}),400
+    usr_password = generate_password_hash(user_password, method='sha256')
+    user = User(user_name,user_email,usr_password)
+    users = db.select_users()
+    for fetch_user in users:
+        if fetch_user[1] == user_name or fetch_user[2] == user_email:
+            return jsonify({'error':'user already exists'}),400
+    user.insert_user()
+    return jsonify({'message':'Account was successfuly created'}),201
+    
+                   
+            
+#Login route for store attendant and admin
 @auths.route('/users/login',methods=['POST'])
 def login():
     """
     Function to login a store attendant into the system
     if the store attendant account doesnot exist, an error is returned
     """
-    try:
-        user_info = request.data
-        login_info = json.loads(user_info)
-        user_name = login_info['name']
-        user_password = str(login_info['password'])
-        if not user_name or not user_password:
-            return jsonify({'error':'username or password cannot be empty'}),400
-        if not isinstance(user_name, str):
-            return jsonify({'error':'username must be a string'}),400
-        if (' ' in user_name) == True:
-            return jsonify({'Error':'user name cannot contain a space'}),400
-        for user in user_db:
-            password = user['user_password']
-            if user['user_name'] == user_name and check_password_hash(password,user_password):
-                user['loggedin'] = True
-                return jsonify({'message':'You are now loggedin'}),200
-        return jsonify({'error':'Wrong username or password'}),400
-    except Exception:
-        return jsonify({'error':'required field/s missing'}), 400  
-
-
-@auths.route('/users/login/admin',methods=['POST'])
-def log_admin():
-    """
-    Function to login an admin into the system
-    if the store attendant account doesnot exist, an error is returned
-
-    """
-    try:
-        admin_data = request.data
-        sign_data = json.loads(admin_data)
-        admin_user_name = sign_data['name']
-        admin_user_password = str(sign_data['password'])
-        if not admin_user_name or not admin_user_password:
-            return jsonify({'error':'username or password cannot be empty'}),400
-        if not isinstance(admin_user_name,str):
-            return jsonify({'error':'username must be a string'}),400
-        if (' ' in admin_user_name) == True:
-            return jsonify({'Error':'user name cannot contain a space'}),400
-        for user in admin_db:
-            password = user['user_password']
-            if user['user_name'] == admin_user_name  and check_password_hash(password,admin_user_password):
-                user['loggedin'] = True
-                return jsonify({'message':'You are now loggedin'}),200
-        return jsonify({'error':'Wrong username or password'}),400
-    except Exception:
-        return jsonify({'error':'required field/s missing'}), 400  
+    user_info = request.data
+    login_info = json.loads(user_info)
+    user_email = login_info['email']
+    user_password = str(login_info['password'])
+    if not user_email or not user_password:
+        return jsonify({'error':'username or password cannot be empty'}),400
+    if not isinstance(user_email, str):
+        return jsonify({'error':'username must be a string'}),400
+    # if (' ' in user_name) == True:
+    #     return jsonify({'Error':'user name cannot contain a space'}),400
+    user_data = db.select_users()
+    for user in user_data:
+        if user[2] == user_email and check_password_hash(user[3],user_password):
+            return jsonify({'message':'You are now loggedin'}),200
+    return jsonify({'error':'Wrong username or password'}),400
+
