+++ app/products/views.py
@@ -1,7 +1,7 @@
 from flask import Blueprint,request,jsonify,json,make_response
 from .models import Product
-from ..auth.utility import admin_required,login_admin_required
-from app.auth.views import admin_db,user_db
+# from ..auth.utility import admin_required,login_admin_required
+# from app.auth.views import admin_db,user_db
